Only set the background after entering vim
This allows vim to guess the value of &background. This will set the
best colors based on system configuration. This will especially help
while using 16 colors: see #50

This disables the fix for morhetz/gruvbox#7 temporarily while entering
vim. It should still be possible to switch to and from light and dark
mode afterwards.