Do not set the def background for todos and errors
Previous settings hardcoded the default background as the background for
todos and errors. This doesn't look good when a todo or an error is at
the cursor line as cursor line highlights change the default background.
What happens is that the todo/error use the background of the rest of a
file and stand out on the cursor-line.